move api calls to modules
